+++ EasyPost/Utilities/Http.cs
@@ -18,6 +18,26 @@ namespace EasyPost.Utilities
         }
 
         /// <summary>
+        ///     Return whether the given status code is a 1xx error.
+        /// </summary>
+        /// <param name="statusCode">Status code to check.</param>
+        /// <returns>True if the status code is a 1xx error, False otherwise.</returns>
+        internal static bool Is1xx(this HttpStatusCode statusCode)
+        {
+            return StatusCodeIs1xx(statusCode);
+        }
+
+        /// <summary>
+        ///     Return whether the given status code is a 2xx error.
+        /// </summary>
+        /// <param name="statusCode">Status code to check.</param>
+        /// <returns>True if the status code is a 2xx error, False otherwise.</returns>
+        internal static bool Is2xx(this HttpStatusCode statusCode)
+        {
+            return StatusCodeIs2xx(statusCode);
+        }
+
+        /// <summary>
         ///     Return whether the given status code is a 3xx error.
         /// </summary>
         /// <param name="statusCode">Status code to check.</param>
@@ -63,7 +83,7 @@ namespace EasyPost.Utilities
         ///     Return whether the given response has an error status code.
         /// </summary>
         /// <param name="response">Response to check.</param>
-        /// <returns>True if the response code is not in the 100-299 range, false otherwise.</returns>
+        /// <returns>True if the response code is not in the 200-299 range, false otherwise.</returns>
         internal static bool ReturnedError(this RestResponse response)
         {
             return !ReturnedNoError(response);
@@ -73,10 +93,10 @@ namespace EasyPost.Utilities
         ///     Return whether the given response has a successful status code.
         /// </summary>
         /// <param name="response">Response to check.</param>
-        /// <returns>True if the response code is in the 100-299 range, false otherwise.</returns>
+        /// <returns>True if the response code is in the 200-299 range, false otherwise.</returns>
         internal static bool ReturnedNoError(this RestResponse response)
         {
-            return StatusCodeBetween(response, 100, 299);
+            return response.StatusCode.Is2xx();
         }
 
         /// <summary>
@@ -113,6 +133,46 @@ namespace EasyPost.Utilities
         internal static bool StatusCodeBetween(RestResponseBase response, int min, int max)
         {
             return StatusCodeBetween(response.StatusCode, min, max);
+        }
+
+        /// <summary>
+        ///     Return whether the given status code is a 1xx error.
+        /// </summary>
+        /// <param name="statusCode">Status code to check.</param>
+        /// <returns>True if the status code is a 1xx error, False otherwise.</returns>
+        internal static bool StatusCodeIs1xx(int statusCode)
+        {
+            return StatusCodeBetween(statusCode, 100, 199);
+        }
+
+        /// <summary>
+        ///     Return whether the given status code is a 1xx error.
+        /// </summary>
+        /// <param name="statusCode">Status code to check.</param>
+        /// <returns>True if the status code is a 1xx error, False otherwise.</returns>
+        internal static bool StatusCodeIs1xx(HttpStatusCode statusCode)
+        {
+            return StatusCodeBetween(statusCode, 100, 199);
+        }
+
+        /// <summary>
+        ///     Return whether the given status code is a 2xx error.
+        /// </summary>
+        /// <param name="statusCode">Status code to check.</param>
+        /// <returns>True if the status code is a 2xx error, False otherwise.</returns>
+        internal static bool StatusCodeIs2xx(int statusCode)
+        {
+            return StatusCodeBetween(statusCode, 200, 299);
+        }
+
+        /// <summary>
+        ///     Return whether the given status code is a 2xx error.
+        /// </summary>
+        /// <param name="statusCode">Status code to check.</param>
+        /// <returns>True if the status code is a 2xx error, False otherwise.</returns>
+        internal static bool StatusCodeIs2xx(HttpStatusCode statusCode)
+        {
+            return StatusCodeBetween(statusCode, 200, 299);
         }
 
         /// <summary>