+++ CHANGELOG.md
@@ -4,9 +4,10 @@
 
 - Improved API error parsing
   - API error message may be an array rather than a string. Arrays will be concatenated (by comma) and returned as a string.
-- Capture 3xx HTTP status codes as errors
+- Capture 1xx and 3xx HTTP status codes as errors
   - Any known 3xx status code from the EasyPost API will throw a `RedirectError` exception
   - Any unknown 3xx status code will throw a `UnexpectedHttpError` exception
+  - Any 1xx status code (known or unknown) will throw a `UnexpectedHttpError` exception
 
 ## v4.0.0-rc1 (2022-09-26)